[Design] friend-tab-navigator
